Liveintent & UID2 ID systems: share EID configuration
